requirements.txt
Django==5.1.2
python-dotenv==1.0.1
requests==2.32.3
SPARQLWrapper==2.0.0